Backport plugin to quartz-monitor
