+++ addRemote.bat
@@ -1,3 +1,3 @@
-call git remote add %1 https://github.com/%1/schwartz-monitor.git
+call git remote add %1 https://github.com/%1/quartz-monitor.git
 call git remote -v
 call git fetch %1
