+++ bintray-setup.sh
@@ -1,0 +1,2 @@
+export BINTRAY_USER=jamescookie
+export BINTRAY_KEY=b88xxxxxxxxx
