+++ grails-app/assets/javascripts/quartz-monitor-main.js
