+++ grails-app/assets/javascripts/quartz-monitor.js
@@ -1,0 +1,1 @@
+//=require quartz-monitor-main
