--- grails-app/assets/javascripts/schwartz-monitor.js
@@ -1,1 +1,0 @@
-//=require schwartz-monitor-main
