+++ grails-app/assets/stylesheets/quartz-monitor.css
@@ -1,0 +1,3 @@
+/*
+*= require quartz-monitor-main
+*/
