--- grails-app/assets/stylesheets/schwartz-monitor.css
@@ -1,3 +1,0 @@
-/*
-*= require schwartz-monitor-main
-*/
