+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'schwartz-monitor'
+rootProject.name = 'quartz-monitor'
