--- src/docs/comparisonWithQuartzMonitorPlugin.adoc
@@ -1,8 +1,0 @@
-[[comparisonWithQuartzMonitorPlugin]]
-== Comparison with the Quartz Monitor Plugin
-
-This plugin was initially based on the quartz-monitor plugin, but have changed during implementation in significant areas.
-
-The Quartz-Monitor plugin uses its own quartzJobFactory, 'QuartzMonitorJobFactory', which is a subclass of the underlying quartz plugins' GrailsJobFactory. This can lead to several problems, e.g. if your application needs to subclass the QuartzJobFactory.
-
-The Schwartz-Monitor plugin uses listeners to monitor the Quartz Scheduler status, so it is basically independent of the underlying Quartz Plugin implementation (in theory, we will see how it works out).
