+++ src/docs/index.tmpl
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
 
-<title>Grails Schwartz Plugin</title>
+<title>Grails Quartz Plugin</title>
 
 <style type="text/css">
 body {
@@ -59,12 +59,12 @@ img {
 </head>
 
 <body>
-	<a href="http://github.com/robertoschwald/grails-schwartz-monitor">
+	<a href="http://github.com/jamescookie/quartz-monitor">
 		<img style="position: absolute; top: 0; right: 0;" src="https://s3.amazonaws.com/github/ribbons/forkme_right_darkblue_121621.png" alt="Fork me on GitHub" />
 	</a>
 
 	<div id="container">
-		<h1>Grails Schwartz Monitor Plugin</h1>
+		<h1>Grails Quartz Monitor Plugin</h1>
 		<table id="properties">
 			<tbody>
 				<tr>
@@ -76,8 +76,8 @@ img {
 					<td>3.0.0 > *</td>
 				</tr>
 				<tr>
-					<td class="propName">Author</td>
-					<td>Robert Oschwald</td>
+					<td class="propName">Authors</td>
+					<td>James Cook, Robert Oschwald</td>
 				</tr>
 			</tbody>
 		</table>
@@ -86,19 +86,19 @@ img {
 
 		<h2>Current Documentation</h2>
 		<ul>
-			<li><a href="/grails-schwartz-monitor/latest/index.html">User guide</a></li>
-			<li><a href="/grails-schwartz-monitor/latest/schwartz-@VERSION@.pdf">User guide PDF</a></li>
-			<li><a href="/grails-schwartz-monitor/latest/schwartz-@VERSION@.epub">User guide EPUB</a></li>
-			<li><a href="/grails-schwartz-monitor/latest/groovydoc/index.html">Groovydoc</a></li>
+			<li><a href="index.html">User guide</a></li>
+			<li><a href="quartz-monitor-@VERSION@.pdf">User guide PDF</a></li>
+			<li><a href="quartz-monitor-@VERSION@.epub">User guide EPUB</a></li>
+			<li><a href="groovydoc/index.html">Groovydoc</a></li>
 		</ul>
 
 		<hr />
 
 		<div class="download">
-			<a href="http://github.com/robertoschwald/grails-schwartz-monitor/zipball/master">
+			<a href="http://github.com/jamescookie/quartz-monitor/zipball/master">
 				<img width="90" src="http://github.com/images/modules/download/zip.png"/>
 			</a>
-			<a href="http://github.com/robertoschwald/grails-schwartz-monitor/tarball/master">
+			<a href="http://github.com/jamescookie/quartz-monitor/tarball/master">
 				<img width="90" src="http://github.com/images/modules/download/tar.png"/>
 			</a>
 		</div>
@@ -106,11 +106,11 @@ img {
 		<h2>Download Source</h2>
 		<p>
 			You can download this project in either
-			<a href="http://github.com/agileorbit/grails-schwartz/zipball/master">zip</a> or
-			<a href="http://github.com/agileorbit/grails-schwartz/tarball/master">tar</a> formats.
+			<a href="http://github.com/jamescookie/quartz-monitor/zipball/master">zip</a> or
+			<a href="http://github.com/jamescookie/quartz-monitor/tarball/master">tar</a> formats.
 		</p>
 		<p>You can also clone the project with <a href="http://git-scm.com">Git</a> by running:
-			<pre>$ git clone git://github.com/robertoschwald/grails-schwartz-monitor</pre>
+			<pre>$ git clone https://github.com/jamescookie/quartz-monitor.git</pre>
 		</p>
 
   </div>
