+++ src/docs/installation.adoc
@@ -1,7 +1,7 @@
 [[installation]]
 == Installation
 
-To use the plugin, you must have the grails-schwartz plugin already installed in your project.
+To use the plugin, you must have the Grails quartz-monitor or grails-schwartz plugin already installed in your project.
 
 === "`Install`" the plugin by adding it to build.gradle
 
@@ -12,7 +12,7 @@ Add a dependency for the plugin by adding it to the `dependencies` block in buil
 ----
 dependencies {
    ...
-   compile 'org.grails.plugins:schwartz-monitor:{project-version}'
+   compile 'org.grails.plugins:quartz-monitor:{project-version}'
    ...
 }
 ----
