+++ src/docs/listeners.adoc
@@ -3,5 +3,5 @@
 
 The plugin uses listeners to monitor the Quartz Scheduler status.
 
-=== grails.plugins.schwartz.listener.QuartzSchedulerListener
+=== grails.plugins.quartz.listener.QuartzSchedulerListener
 Main Listener to monitor Scheduler configuration, runs, etc.
