+++ travis-build.sh
@@ -16,7 +16,7 @@ if [[ $EXIT_STATUS -eq 0 ]]; then
 fi
 
 if [[ -n $TRAVIS_TAG && $TRAVIS_PULL_REQUEST == 'false' && $EXIT_STATUS -eq 0 ]]; then
-    echo "Uploading to bintray"
+  echo "Uploading to bintray"
 	./gradlew bintrayUpload --stacktrace
 
 	./gradlew docs --stacktrace
@@ -25,7 +25,7 @@ if [[ -n $TRAVIS_TAG && $TRAVIS_PULL_REQUEST == 'false' && $EXIT_STATUS -eq 0 ]]
 	git config --global credential.helper "store --file=~/.git-credentials"
 	echo "https://$GH_TOKEN:@github.com" > ~/.git-credentials
 
-    echo "Cloning gh-pages branch from $TRAVIS_REPO_SLUG"
+  echo "Cloning gh-pages branch from $TRAVIS_REPO_SLUG"
 	git clone https://${GH_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git -b gh-pages gh-pages --single-branch > /dev/null
 	cd gh-pages
